perf(build): avoid vercel re-linking on subsequent builds
For the first version of an app, the Vercel project is linked as usual.
For subsequent versions, the worker now attempts to reuse the existing
`.vercel` folder, which is downloaded from S3 along with the rest of
the code. This prevents unnecessary Vercel API calls and speeds up
the build process. The `.vercel` folder is no longer excluded from
the packaged code.